+++ source/js/modules/rules.js
@@ -1,4 +1,4 @@
-import {AccentTypographyBuild} from "../utils/helpers/animation.helpers";
+import {AccentTypographyBuild} from "../utils/helpers/animation";
 
 export default () => {
   const animationRulesTitle = new AccentTypographyBuild(
