+++ source/js/modules/slider.js
@@ -1,5 +1,5 @@
 import Swiper from "swiper";
-import {AccentTypographyBuild} from "../utils/helpers/animation.helpers";
+import {AccentTypographyBuild} from "../utils/helpers/animation";
 
 export default () => {
   // slider
